Move IVal to inner class of ActivityBased and change its name to ValueActivity. + Change threshold to switch between Array and Map + remove call to `ensureCapacity` from `get`
